services/collector/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
kafka-python==2.0.2
pydantic==2.5.0
python-dotenv==1.0.0
python-multipart==0.0.6
prometheus-client==0.19.0